gcp-python/requirements.txt
google.cloud.storage
google.cloud.bigquery
pandas
fiona
geopandas
requests
